+++ fixtures/decorators/tests/test_generated_bindings.rs
@@ -1,0 +1,4 @@
+uniffi_macros::build_foreign_language_testcases!(
+    ["src/decorators.udl"],
+    ["tests/bindings/test_decorators.kts"]
+);