+++ uniffi_bindgen/src/bindings/kotlin/mod.rs
@@ -51,7 +51,12 @@ fn full_bindings_path(config: &Config, out_dir: &Path) -> Result<PathBuf> {
 
 /// Generate kotlin bindings for the given namespace, then use the kotlin
 /// command-line tools to compile them into a .jar file.
-pub fn compile_bindings(config: &Config, ci: &ComponentInterface, out_dir: &Path) -> Result<()> {
+pub fn compile_bindings(
+    config: &Config,
+    ci: &ComponentInterface,
+    out_dir: &Path,
+    extra_files: &[PathBuf],
+) -> Result<()> {
     let mut kt_file = full_bindings_path(config, out_dir)?;
     kt_file.push(format!("{}.kt", ci.namespace()));
     let mut jar_file = PathBuf::from(out_dir);
@@ -63,6 +68,7 @@ pub fn compile_bindings(config: &Config, ci: &ComponentInterface, out_dir: &Path
         .arg("-classpath")
         .arg(env::var("CLASSPATH").unwrap_or_else(|_| "".to_string()))
         .arg(&kt_file)
+        .args(extra_files)
         .arg("-d")
         .arg(jar_file)
         .spawn()