+++ uniffi_bindgen/src/interface/attributes.rs
@@ -36,6 +36,9 @@ pub(super) enum Attribute {
     External(String),
     // Custom type on the scaffolding side
     Custom,
+    // Used to specify decorators for interfaces
+    WithDecoratedClass(String),
+    WithDecoratorMethod(String),
 }
 
 impl Attribute {
@@ -71,6 +74,12 @@ impl TryFrom<&weedle::attribute::ExtendedAttribute<'_>> for Attribute {
                     "Throws" => Ok(Attribute::Throws(name_from_id_or_string(&identity.rhs))),
                     "Self" => Ok(Attribute::SelfType(SelfType::try_from(&identity.rhs)?)),
                     "External" => Ok(Attribute::External(name_from_id_or_string(&identity.rhs))),
+                    "Decorated" => Ok(Attribute::WithDecoratedClass(name_from_id_or_string(
+                        &identity.rhs,
+                    ))),
+                    "CallsWith" => Ok(Attribute::WithDecoratorMethod(name_from_id_or_string(
+                        &identity.rhs,
+                    ))),
                     _ => anyhow::bail!(
                         "Attribute identity Identifier not supported: {:?}",
                         identity.lhs_identifier.0
@@ -252,6 +261,15 @@ impl InterfaceAttributes {
             .iter()
             .any(|attr| matches!(attr, Attribute::Threadsafe))
     }
+
+    pub(super) fn get_decorated_class(&self) -> Option<&str> {
+        self.0.iter().find_map(|attr| match attr {
+            // This will hopefully return a helpful compilation error
+            // if the decorator method is not defined.
+            Attribute::WithDecoratedClass(inner) => Some(inner.as_ref()),
+            _ => None,
+        })
+    }
 }
 
 impl TryFrom<&weedle::attribute::ExtendedAttributeList<'_>> for InterfaceAttributes {
@@ -263,6 +281,7 @@ impl TryFrom<&weedle::attribute::ExtendedAttributeList<'_>> for InterfaceAttribu
             Attribute::Enum => Ok(()),
             Attribute::Error => Ok(()),
             Attribute::Threadsafe => Ok(()),
+            Attribute::WithDecoratedClass(_) => Ok(()),
             _ => bail!(format!("{:?} not supported for interface definition", attr)),
         })?;
         // Can't be both `[Threadsafe]` and an `[Enum]`.
@@ -346,6 +365,15 @@ impl MethodAttributes {
             .iter()
             .any(|attr| matches!(attr, Attribute::SelfType(SelfType::ByArc)))
     }
+
+    pub fn get_decorator_method(&self) -> Option<&str> {
+        self.0.iter().find_map(|attr| match attr {
+            // This will hopefully return a helpful compilation error
+            // if the decorator method is not defined.
+            Attribute::WithDecoratorMethod(inner) => Some(inner.as_ref()),
+            _ => None,
+        })
+    }
 }
 
 impl TryFrom<&weedle::attribute::ExtendedAttributeList<'_>> for MethodAttributes {
@@ -356,6 +384,7 @@ impl TryFrom<&weedle::attribute::ExtendedAttributeList<'_>> for MethodAttributes
         let attrs = parse_attributes(weedle_attributes, |attr| match attr {
             Attribute::SelfType(_) => Ok(()),
             Attribute::Throws(_) => Ok(()),
+            &Attribute::WithDecoratorMethod(_) => Ok(()),
             _ => bail!(format!("{:?} not supported for methods", attr)),
         })?;
         Ok(Self(attrs))
@@ -529,6 +558,20 @@ mod test {
     }
 
     #[test]
+    fn test_decorator() -> Result<()> {
+        let (_, node) =
+            weedle::attribute::ExtendedAttribute::parse("Decorated=MyDecoratedClass").unwrap();
+        let attr = Attribute::try_from(&node)?;
+        assert!(matches!(attr, Attribute::WithDecoratedClass(nm) if nm == "MyDecoratedClass"));
+
+        let (_, node) = weedle::attribute::ExtendedAttribute::parse("CallsWith=asyncCall").unwrap();
+        let attr = Attribute::try_from(&node)?;
+        assert!(matches!(attr, Attribute::WithDecoratorMethod(nm) if nm == "asyncCall"));
+
+        Ok(())
+    }
+
+    #[test]
     fn test_unsupported() {
         let (_, node) =
             weedle::attribute::ExtendedAttribute::parse("UnsupportedAttribute").unwrap();
@@ -579,6 +622,14 @@ mod test {
             err.to_string(),
             "SelfType(ByArc) not supported for functions"
         );
+
+        let (_, node) =
+            weedle::attribute::ExtendedAttributeList::parse("[CallsWith=asyncDispatch]").unwrap();
+        let err = FunctionAttributes::try_from(&node).unwrap_err();
+        assert_eq!(
+            err.to_string(),
+            "WithDecoratorMethod(\"asyncDispatch\") not supported for functions"
+        );
     }
 
     #[test]
@@ -592,17 +643,28 @@ mod test {
         let attrs = MethodAttributes::try_from(&node).unwrap();
         assert!(!attrs.get_self_by_arc());
         assert!(attrs.get_throws_err().is_none());
+        assert!(attrs.get_decorator_method().is_none());
 
         let (_, node) =
             weedle::attribute::ExtendedAttributeList::parse("[Self=ByArc, Throws=Error]").unwrap();
         let attrs = MethodAttributes::try_from(&node).unwrap();
         assert!(attrs.get_self_by_arc());
         assert!(attrs.get_throws_err().is_some());
+        assert!(attrs.get_decorator_method().is_none());
 
         let (_, node) = weedle::attribute::ExtendedAttributeList::parse("[Self=ByArc]").unwrap();
         let attrs = MethodAttributes::try_from(&node).unwrap();
         assert!(attrs.get_self_by_arc());
         assert!(attrs.get_throws_err().is_none());
+        assert!(attrs.get_decorator_method().is_none());
+
+        let (_, node) =
+            weedle::attribute::ExtendedAttributeList::parse("[CallsWith=asyncDispatch]").unwrap();
+        let attrs = MethodAttributes::try_from(&node).unwrap();
+        assert!(matches!(
+            attrs.get_decorator_method(),
+            Some("asyncDispatch")
+        ));
     }
 
     #[test]